requirements.txt
Flask
python-dotenv
pymongo
waitress